requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
multidict==6.0.5
packaging==24.1
pi==0.1.2
psycopg2-binary==2.9.9
requests==2.32.3
simplejson==3.19.2
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
